perf: Only update Transform if it actually changed
Fixes some nasty performance given that `PreventChildScaling` is called in `_PhysicsProcess`. Apparently, the cost of setting Transform grows with the number of descendents to a Node, even if the Transform is identical to what it was before. So, when a lot of conveyor legs are present, the perf hit is relatively huge.

And if I'm wrong about that, find some other reason this code fixes the problem that varies with the length, not the number of assemblies.