Move RollerCorner material management into CurvedRollerConveyor
RollerCorner animation was moving way too fast. The reason was that the Material was shared across all RollerCorners and all 24 RollerCorners were each updating it. Instead of shifting one step per tick, the material shifted 24.

Now, the CurvedRollerConveyor does the updating so it only happens once. Also, to prevent other CurvedRollerConveyor instances from interfering with the material, it is now Duplicated to make it unique to each CurvedRollerConveyor and applied as a material override on all the RollerCorner instances.