Fix initialization when default scene property values don't match the script's
For example, BeltConveyorAssembly.tscn has a default length of 4 despite having a base length of 1. Somehow, `Transform` is already initialized with 4 before the constructor is called and our signal handlers are setup. Perhaps our script is attached after `Transform` has been set? Anyway, we trigger our SetTransform handler manually because it was skipped. It triggers the remaining signals and handlers as needed.