Fix regression in ConveyorAssembly preview node
PreventScaling was called in the constructor (before parent exists). This caused ConstrainApparentScale to be called with the assumption that there was no parent scale. This caused the scale from the PackedScene (which was already correct) to be reset to (1, 1, 1), which was incorrect once the parent was present.

Maybe we should also listen to `NOTIFICATION_PARENTED`?