Add AsNode3D method to IConveyor
This is a workaround for the fact that we can't declare a requirement to inherit from Node3D. All IConveyors that we create have extended Node3D, and it's common to write logic that depends on this fact. Ideally Godot would provide an INode3D interface implemented by Node3D, and we could depend on it, but it doesn't.

There are third-party libraries that add interfaces of all Godot classes, but they're kinda kludgey (because they too can't modify the built-in classes to actually implement the new interfaces for them). I don't want to integrate one of those solutions throughout the project just to fix this one interface.