Fix failure to unload .NET assemblies
Occasionally, rebuilding the project would display the error message ".NET: Failed to unload assemblies." and the project would need to be reloaded to continue.

Apparently, Godot was trying to serialize a delegate subscribed to the `OnSelectionChanged` event. The target of the delegate was a method on Main, a Node instance which had been disposed. When Godot looked up Main's instance id, the lookup method discovered that Main had been disposed and threw an `ObjectDisposedException`. This caused the entire assembly reload to fail.

It seems like throwing the exception may have been a preventative measure. In any case, the whole problem can be avoided by not storing a delegate at all.

This commit replaces the event subscription with an equivalent Godot signal connection. Instead of a delegate, a Godot Callable is used. The syntax is uglier, but it lets Godot manage disconnecting the signal for us.

I found more info on this problem here: https://docs.godotengine.org/en/latest/tutorials/scripting/c_sharp/c_sharp_signals.html