Deprecate ConveyorAutomaticLength
It's now hidden, but still functions the same way. This way, existing scenes that have it enabled aren't changed. However, no one will want to turn it on in the first place.

The property will still be visible when it's not set to its default value. This allows users to reset it.