Improve `SpurConveyor` edge position formula
When adjusting `angle_downstream` or `angle_upstream`, pivot the line of contact around a fixed point at z=0. This is a line that conveyors will always touch but never cross.

The benefit of this change is that it's now possible to align the midpoint of a `SpurConveyor`'s edge with another conveyor, then adjust the angle or `conveyor_count` afterwards without further changes to `length`.